create kbb project structure for automation
- create kbb project structure for automation
- create BasePage class
- create BaseTest class
