--- demo/src/test/java/com/example/test/pages/ProductPage.java
@@ -1,26 +1,0 @@
-package com.example.test.pages;
-
-import org.openqa.selenium.By;
-import org.openqa.selenium.WebDriver;
-import org.openqa.selenium.WebElement;
-import org.openqa.selenium.support.ui.ExpectedConditions;
-
-public class ProductPage extends BasePage {
-
-    public ProductPage(WebDriver driver) {
-        super(driver);
-    }
-    //#region Declare Web Object
-    private String sleeveProduct = "//a[@class='product_img_link'] //img[@alt='Faded Short Sleeve T-shirts']";
-    private WebElement shortSleeveProduct = wait.until(ExpectedConditions.presenceOfElementLocated(By.xpath(sleeveProduct)));
-    
-    //#endregion
-
-    //#region page actions 
-    public void clickOnProductElement(){
-        this.shortSleeveProduct.click();
-    }
-    //#endregion
-
-    
-}
